+++ app/Http/Controllers/Admin/ActivityPolicyController.php
@@ -29,19 +29,19 @@ class ActivityPolicyController extends Controller
 
     public function index($activity_id)
     {
-        // abort_if(Gate::denies('activity_policy_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('activity_policy_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $activity = $this->activity_service->getById($activity_id);
         return view('activity_policy.index', compact('activity'));
     }
 
     public function getData(Request $request)
     {
-        // abort_if(Gate::denies('activity_policy_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('activity_policy_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return $this->activity_policy_service->getSource($request->all());
     }
     public function store(Request $request)
     {
-        // abort_if(Gate::denies('activity_policy_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('activity_policy_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $validation = Validator::make(
             $request->all(),
             [
@@ -81,7 +81,7 @@ class ActivityPolicyController extends Controller
     }
     public function destroy($id)
     {
-        // abort_if(Gate::denies('activity_policy_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('activity_policy_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             $activity_policy = $this->activity_policy_service->deleteById($id);
             return $this->success(
